Build 2.0.0.39
---"Vanguard: Shield Specialization V" now correctly has "Vanguard: Shield Specialization IV" as a prerequisite (Reported by Deathbringer)
---Spell DCs no longer have an additional +10 to their totals (Reported by christhemiss)
---Some class descriptions updated to match that the class is now Free to Play.
---Myth Drannor Favor rewards tiers set as per https://www.ddo.com/update-notes/ddo-update-69-release-notes
---Filigree tooltips will now show the correct number of set bonus stacks
---Augments that grant set bonuses will now show the correct number of set bonus stacks in tooltips
---New option during item select to "Ignore Minor Artifacts"
---The following new feats now have their correct images:
------Arcane Actions
------Channel Power
------Face Me
------Guarded Actions
------Intimidating Shout (Renamed from Reverberating Shout), description also updated
------Defensive Tumble (Description updated)
------Inspiring Bond (Description updated)
------Paranoid
------Patience
------Trick Shot
------Titan's Blood (Provided by Brakkart)
---Quest added to the quest list:
------Voices of the Dead
---Correct base favor values set for quests:
------The Safehold (5)
------Death Hosts this Banquet (5)
------Secrets of the Red Wizards (6)
------The House of Gems (5)
------Ogmha's Song (5)
------The Zulkir's Plan (6)
------The Key to the Mythal (7)
------Book Burning (4)
------Ettercap Incursion (5)
------Portal to Below (5)
------Seeds of Decay (5)
------Voices of the Dead (5)
------Times Long Past (7) (Info from Duncanthrax)

---Downloaded and parsed new Wiki item data (307 new items)
---NOTE THAT ITEM DOWNLOADED FROM WIKI ARE LAMANNIA VERSIONS AND MAY NOT MATCH LIVE, TILL WIKI CATCHES UP
---Updated wiki item parser to handle new effects from U69 items
---Manually fixed ~12 items
---Item images not yet set as not available on wiki yet
---Many item has TBD flavor text and are missing drop locations

+++ DDOBuilder/Spell.cpp
@@ -348,7 +348,7 @@ int Spell::TotalCost(const Build& build) const
         BreakdownItem *pBI = FindBreakdown(Breakdown_Accelerate);
         if (pBI != NULL)
         {
-            cost += pBI->Total();
+            cost += static_cast<int>(pBI->Total());
         }
     }
     if (HasEmbolden() && build.IsStanceActive("Embolden Spell"))
@@ -356,7 +356,7 @@ int Spell::TotalCost(const Build& build) const
         BreakdownItem* pBI = FindBreakdown(Breakdown_Embolden);
         if (pBI != NULL)
         {
-            cost += pBI->Total();
+            cost += static_cast<int>(pBI->Total());
         }
     }
     if (HasEmpower() && build.IsStanceActive("Empower Spell"))
@@ -364,7 +364,7 @@ int Spell::TotalCost(const Build& build) const
         BreakdownItem* pBI = FindBreakdown(Breakdown_Empower);
         if (pBI != NULL)
         {
-            cost += pBI->Total();
+            cost += static_cast<int>(pBI->Total());
         }
     }
     if (HasEmpowerHealing() && build.IsStanceActive("Empower Healing Spell"))
@@ -372,7 +372,7 @@ int Spell::TotalCost(const Build& build) const
         BreakdownItem* pBI = FindBreakdown(Breakdown_EmpowerHealing);
         if (pBI != NULL)
         {
-            cost += pBI->Total();
+            cost += static_cast<int>(pBI->Total());
         }
     }
     if (HasEnlarge() && build.IsStanceActive("Enlarge Spell"))
@@ -380,7 +380,7 @@ int Spell::TotalCost(const Build& build) const
         BreakdownItem* pBI = FindBreakdown(Breakdown_Enlarge);
         if (pBI != NULL)
         {
-            cost += pBI->Total();
+            cost += static_cast<int>(pBI->Total());
         }
     }
     if (HasExtend() && build.IsStanceActive("Extend Spell"))
@@ -388,7 +388,7 @@ int Spell::TotalCost(const Build& build) const
         BreakdownItem* pBI = FindBreakdown(Breakdown_Extend);
         if (pBI != NULL)
         {
-            cost += pBI->Total();
+            cost += static_cast<int>(pBI->Total());
         }
     }
     if (HasHeighten() && build.IsStanceActive("Heighten Spell"))
@@ -402,7 +402,7 @@ int Spell::TotalCost(const Build& build) const
             size_t heightenedLevels = maxLevel - Level();
             if (heightenedLevels > 0)
             {
-                cost += (pBI->Total() * heightenedLevels);
+                cost += (static_cast<int>(pBI->Total()) * heightenedLevels);
             }
         }
     }
@@ -411,7 +411,7 @@ int Spell::TotalCost(const Build& build) const
         BreakdownItem* pBI = FindBreakdown(Breakdown_Intensify);
         if (pBI != NULL)
         {
-            cost += pBI->Total();
+            cost += static_cast<int>(pBI->Total());
         }
     }
     if (HasMaximize() && build.IsStanceActive("Maximize Spell"))
@@ -419,7 +419,7 @@ int Spell::TotalCost(const Build& build) const
         BreakdownItem* pBI = FindBreakdown(Breakdown_Maximize);
         if (pBI != NULL)
         {
-            cost += pBI->Total();
+            cost += static_cast<int>(pBI->Total());
         }
     }
     if (HasQuicken() && build.IsStanceActive("Quicken Spell"))
@@ -427,7 +427,7 @@ int Spell::TotalCost(const Build& build) const
         BreakdownItem* pBI = FindBreakdown(Breakdown_Quicken);
         if (pBI != NULL)
         {
-            cost += pBI->Total();
+            cost += static_cast<int>(pBI->Total());
         }
     }
     return cost;
